requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
hydra-core>=1.3.0
wandb>=0.15.0
tensorboard>=2.13.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
albumentations>=1.3.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Data & Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Compression & Entropy Coding
# compressai>=1.2.0
# range-coder>=1.0.0
# arithmetic-coding>=1.0.0

# Model Optimization
onnx>=1.14.0
tensorrt>=8.6.0
torch-tensorrt>=1.4.0

# Utilities
tqdm>=4.65.0
omegaconf>=2.3.0
rich>=13.0.0
typer>=0.9.0
psutil>=5.9.0
GPUtil>=1.4.0

# Development & Testing
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter & Analysis
jupyter>=1.0.0
ipywidgets>=8.0.0
plotly>=5.15.0